some changes for decomposition
