+++ .eslintrc
@@ -19,7 +19,8 @@
   },
   "rules": {
     "linebreak-style": 0,
-    "no-param-reassign": 0
+    "no-param-reassign": 0,
+    "consistent-return": 0
   },
   "parserOptions": {
     "sourceType": "module",
