--- src/modules/jquery.module.js
@@ -1,5 +1,0 @@
-import * as jQuery from 'jquery';
-
-window.jQuery = jQuery;
-
-export default window.jQuery.noConflict(true);
